chore: add types file for table folder
