+++ apps/hubble-stats/components/ShieldedTables/index.ts
@@ -1,3 +1,1 @@
 export * from './ShieldedTables';
-
-export * from './types.d';
