+++ apps/hubble-stats/components/table/PoolTypeCell.tsx
@@ -2,12 +2,7 @@ import { FC } from 'react';
 import cx from 'classnames';
 import { Chip, ChipProps } from '@webb-tools/webb-ui-components';
 
-import { PoolType } from '../ShieldedTables';
-
-export interface PoolTypeCellProps {
-  type: PoolType;
-  className?: string;
-}
+import { PoolTypeCellProps } from './types';
 
 const typeColorMap: Record<PoolTypeCellProps['type'], ChipProps['color']> = {
   single: 'purple',
